requirements.txt
altgraph==0.17.2
astroid==2.11.2
auto-py-to-exe==2.18.0
bottle==0.12.19
bottle-websocket==0.2.9
cffi==1.15.0
chardet==4.0.0
click==8.1.2
colorama==0.4.4
cryptography==36.0.2
dill==0.3.4
Eel==0.12.4
et-xmlfile==1.1.0
future==0.18.2
gevent==21.12.0
gevent-websocket==0.10.1
greenlet==1.1.2
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.7.0
openpyxl==3.0.9
pdfminer.six==20220319
pefile==2021.9.3
pep517==0.12.0
pip-tools==6.5.1
platformdirs==2.5.1
pycparser==2.21
pyinstaller==4.10
pyinstaller-hooks-contrib==2022.3
pylint==2.13.4
PyPDF2==1.26.0
pywin32-ctypes==0.2.0
tomli==2.0.1
typing_extensions==4.1.1
whichcraft==0.6.1
wrapt==1.14.0
zope.event==4.5.0
zope.interface==5.4.0
